From sysadmin.html in cyrus-sasl2.rpm (Suse Linux 8.2):
-----
Authentication and authorization identifiers

An important concept to become familiar with is the difference between an
"authorization identifier" and an "authentication identifier".

userid (user id, authorization id)
    The userid is the identifier an application uses to check allowable
    options. On my Unix system, the user "bovik" (the account of Harry Q.
    Bovik) is allowed to write to "/home/bovik" and its subdirectories but not
    to "/etc".

authid (authentication id)
    The authentication identifier is the identifier that is being checked.
    "bovik"'s password might be "qqqq", and the system will authenticate anyone
    who knows "qqqq" as "bovik". However, it's possible to authenticate as one
    user but act as another user. For instance, Harry might be away on vacation
    and assign one of his graduate students, Jane, to read his mail. He might
    then allow Jane to act as him merely by supplying her password and her id
    as authentication but requesting authorization as "bovik". So Jane might
    log in with an authentication identifier of "jane" and an authorization id
    of "bovik" and her own (Jane's) password.

Applications can set their own proxy policies; by default, the SASL library
will only allow the same user to act for another (that is, userid must equal
authid).
------

The attached patch will allow an Administrator knowing only the admin
password to install a sieve script on behalf of any user. The patch is
incomplete as in that it only works with PLAIN authentication (this
is what I needed in my project). It can be easily extended for other
AUTH MECHS.